fixed tracker

+++ AI-HSE/Sound/HW_3/src/metrics/tracker.py
@@ -16,7 +16,9 @@ class MetricTracker:
                 from each batch.
         """
         self.writer = writer
-        self._data = pd.DataFrame(index=keys, columns=["total", "counts", "average"])
+        self._data = pd.DataFrame(
+            index=keys, columns=["total", "counts", "average"], dtype=float
+        )
         self.reset()
 
     def reset(self):
